requirements.txt
streamlit
flask  
flask-cors
google-generativeai
pdfplumber 
python-dotenv
plotly
pandas
pillow
